requirements.txt
playwright
pandas
beautifulsoup4
geopy
python-dotenv
fake-useragent
playwright-stealth

pywhatkit
